+++ CHANGES.md
@@ -1,7 +1,8 @@
 # pyfakefs Release Notes
 The released versions correspond to PyPi releases.
 
-## Version 4.6.0 (as yet unreleased)
+## [Version 4.5.2](https://pypi.python.org/pypi/pyfakefs/4.5.2) (2021-11-07)
+This is a bugfix release.
 
 ### Changes
 * `os.listdir`, `os.scandir` and `pathlib.Path.listdir` now return the